renamed manticore_evm to evm_opcodes
